+++ tests/testthat/test-reporter-check.R
@@ -1,5 +1,5 @@
 test_that("basic report works", {
-  on.exit(unlink(test_path("testthat-problems.rds")))
+  withr::defer(file.remove(test_path("testthat-problems.rds")))
   expect_snapshot_reporter(CheckReporter$new())
 
   rds <- test_path("testthat-problems.rds")
@@ -13,7 +13,16 @@ test_that("doesn't truncate long lines", {
 })
 
 test_that("always shows summary", {
-  on.exit(unlink(test_path("testthat-problems.rds")))
+  file.create(test_path("testthat-problems.rds"))
 
   expect_snapshot_reporter(CheckReporter$new(), test_path("reporters/successes.R"))
+  # and cleans up testthat-problems
+  expect_false(file.exists(test_path("testthat-problems.rds")))
+})
+
+test_that("shows warnings when not on CRAN", {
+  on.exit(unlink(test_path("testthat-problems.rds")))
+
+  withr::local_options("NOT_CRAN" = "true")
+  expect_snapshot_reporter(CheckReporter$new(), test_path("reporters/tests.R"))
 })